// src/app/intefaces/resume-form.interface.ts
export interface ResumeForm {
  professionName: string;
  firstName: string;
  lastName: string;
  middleName: string;
  isMale: boolean;
  dayOfBirth: string;
  monthOfBirth: string;
  yearOfBirth: string;
  email: string;
  phoneNumber: string;
  educationId: string;
  specialization: string
  aboutMe: string;
}