// src/app/intefaces/resume-forms-values-to-update.interface.ts
export interface ResumeFormValues {
    firstName?: string | null | undefined;
    lastName?: string | null | undefined;
    middleName?: string | null | undefined;
    dayOfBirth?: string | null | undefined;
    monthOfBirth?: number | null | undefined;
    yearOfBirth?: string | null | undefined;
    isMale?: boolean | null | undefined;
    phoneNumber?: string | null | undefined;
    email?: string | null | undefined;
    professionName?: string | null | undefined;
    educationId?: number | null | undefined;
    specialization?: string | null | undefined;
    aboutMe?: string | null | undefined;
}